enemy spawning is now based on polynomials/sine waves
each enemy has it's own, to determine start and end times for it's spawning, and how that value changes. some use sine waves (such as the more unique bombers/martyrs and loot goblins), though mayhaps having these be somewhat bound to a quadratic so they can be phased out for more powerful versions of themselves could be good?